+++ CHANGELOG.md
@@ -19,6 +19,10 @@ and this project adheres to
 
 ### Deprecated
 
+<!--
+    TODO(a.garipov): Remove this deprecation, if v0.108.0 is released before
+    the Go 1.18 release.
+-->
 - Go 1.17 support.  v0.109.0 will require at least Go 1.18 to build.
 
 ### Removed
@@ -36,10 +40,13 @@ and this project adheres to
 
 ### Fixed
 
+- Service not being stopped before running the `uninstall` service action
+  ([#3868]).
 - Legacy DNS rewrites responding from upstream when a request other than `A` or
   `AAAA` is received ([#4008]).
 - Panic on port availability check during installation ([#3987]).
 
+[#3868]: https://github.com/AdguardTeam/AdGuardHome/issues/3868
 [#3987]: https://github.com/AdguardTeam/AdGuardHome/issues/3987
 [#4008]: https://github.com/AdguardTeam/AdGuardHome/issues/4008